initial generic restructuring
