+++ src/main/scala/ai/privado/passes/PropertyEnvLinkerPass.scala
@@ -1,54 +1,47 @@
-/*
- * This file is part of Privado OSS.
- *
- * Privado is an open source static code analysis tool to discover data flows in the code.
- * Copyright (C) 2022 Privado, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * For more information, contact support@privado.ai
- *
- */
-
-package ai.privado.languageEngine.java.passes.config
+package ai.privado.passes
 
 import ai.privado.languageEngine.java.language.NodeStarters
+import ai.privado.model.Language
 import ai.privado.tagger.PrivadoParallelCpgPass
-import io.shiftleft.codepropertygraph.generated.nodes.*
 import io.shiftleft.codepropertygraph.generated.{Cpg, EdgeTypes}
+import io.shiftleft.codepropertygraph.generated.nodes.JavaProperty
+import io.shiftleft.codepropertygraph.generated.nodes.*
 import io.shiftleft.semanticcpg.language.*
 import org.slf4j.LoggerFactory
 import overflowdb.BatchedUpdate
 
-/** This pass creates a graph layer for Java `.properties` files.
-  */
-class JavaPropertyLinkerPass(cpg: Cpg) extends PrivadoParallelCpgPass[JavaProperty](cpg) {
+/** This pass creates a graph layer for env call from property node
+ */
+
+class PropertyEnvLinkerPass(cpg: Cpg, language: Language.Value) extends PrivadoParallelCpgPass[JavaProperty](cpg) {
 
   implicit val resolver: NoResolve.type = NoResolve
-  val logger                            = LoggerFactory.getLogger(getClass)
+  val logger = LoggerFactory.getLogger(getClass)
 
   override def generateParts(): Array[_ <: AnyRef] = {
-    cpg.property.iterator.filter(pair => pair.name.nonEmpty && pair.value.nonEmpty).toArray
+    cpg.property
+      .iterator
+      .filter(pair => pair.name.nonEmpty && pair.value.nonEmpty)
+      .toArray
   }
 
   override def runOnPart(builder: DiffGraphBuilder, property: JavaProperty): Unit = {
-    connectProperties(property, builder)
+    language match
+      case Language.JAVA => {
+        connectEnvProperty(javaMatchingEnvLiteralToPropertyNode(property.name), property, builder)
+        connectAnnotatedParameters(property, builder)
+      }
+      case Language.GO => {
+        connectEnvProperty(goMatchingEnvLiteralsToPropertyNode(property.name), property, builder)
+      }
   }
 
-  private def connectProperties(property: JavaProperty, builder: DiffGraphBuilder): Unit = {
-    connectGetPropertyLiterals(property, builder)
-    connectAnnotatedParameters(property, builder)
+  // This method used to connect the property node with literal where it used
+  private def connectEnvProperty(literalNodes: List[Literal], propertyNode: JavaProperty, builder: DiffGraphBuilder): Unit = {
+    literalNodes.foreach(lit => {
+      builder.addEdge(propertyNode, lit, EdgeTypes.IS_USED_AT)
+      builder.addEdge(lit, propertyNode, EdgeTypes.ORIGINAL_PROPERTY)
+    })
   }
 
   private def connectAnnotatedParameters(propertyNode: JavaProperty, builder: BatchedUpdate.DiffGraphBuilder): Unit = {
@@ -88,7 +81,7 @@ class JavaPropertyLinkerPass(cpg: Cpg) extends PrivadoParallelCpgPass[JavaProper
   }
 
   /** List of all methods annotated with Spring's `Value` annotation, along with the method node
-    */
+   */
   private def annotatedMethods(): List[(AnnotationParameterAssign, Method)] = cpg.annotation
     .nameExact("Value")
     .filter(_.method.nonEmpty)
@@ -97,14 +90,14 @@ class JavaPropertyLinkerPass(cpg: Cpg) extends PrivadoParallelCpgPass[JavaProper
     .toList
 
   /** List of all parameters annotated with Spring's `Value` annotation, along with the property name.
-    */
+   */
   private def annotatedParameters(): List[(MethodParameterIn, String)] = cpg.annotation
     .nameExact("Value")
     .filter(_.parameter.nonEmpty)
     .filter(_.parameterAssign.code("\\\"\\$\\{.*\\}\\\"").nonEmpty)
     .map { x =>
       val literalName = x.parameterAssign.code.next()
-      val value       = Option(literalName.slice(3, literalName.length - 2)).getOrElse("")
+      val value = Option(literalName.slice(3, literalName.length - 2)).getOrElse("")
       (x.parameter.next(), value)
     }
     .filter { (_, value) =>
@@ -113,7 +106,7 @@ class JavaPropertyLinkerPass(cpg: Cpg) extends PrivadoParallelCpgPass[JavaProper
     .toList
 
   /** List of all members annotated with Spring's `Value` annotation, along with the property name.
-    */
+   */
   private def annotatedMembers(): List[(AnnotationParameterAssign, Member)] = cpg.annotation
     .nameExact("Value")
     .filter(_.member.nonEmpty)
@@ -121,18 +114,24 @@ class JavaPropertyLinkerPass(cpg: Cpg) extends PrivadoParallelCpgPass[JavaProper
     .map { x => (x.parameterAssign.next(), x.member.next()) }
     .toList
 
-  /** In this method, we attempt to identify users of properties and connect them to property nodes.
-    */
-  private def connectGetPropertyLiterals(propertyNode: JavaProperty, builder: BatchedUpdate.DiffGraphBuilder): Unit = {
-    matchingLiteralsInGetPropertyCalls(propertyNode.name).foreach { lit =>
-      builder.addEdge(propertyNode, lit, EdgeTypes.IS_USED_AT)
-      builder.addEdge(lit, propertyNode, EdgeTypes.ORIGINAL_PROPERTY)
-    }
+  // List all literal fetching property value for GO
+  // ex: os.get("KEY")
+  private def goMatchingEnvLiteralsToPropertyNode(propertyName: String): List[Literal] = {
+    //To get every dot seperated part of the env key where each element is progressively longer,
+    // containing one more dot-separated value
+    val parts = propertyName.split("\\.")
+    val propertyKey = (1 to parts.length).map(i => parts.take(i).mkString(".")).mkString("|")
+    cpg.literal
+      .codeExact("\"" + propertyKey + "\"")
+      .filter(_.inCall.name("(?i).*getenv").nonEmpty)
+      .toList
   }
 
-  private def matchingLiteralsInGetPropertyCalls(propertyName: String): List[Literal] = cpg.literal
+  // List all literal fetching property value for JAVA
+  // ex: System.getProperty("KEY")
+  private def javaMatchingEnvLiteralToPropertyNode(propertyName: String): List[Literal]
+  = cpg.literal
     .codeExact("\"" + propertyName + "\"")
     .filter(_.inCall.name(".*getProperty").nonEmpty)
     .toList
-
 }
