+++ src/test/scala/ai/privado/passes/propertyEnv/JavaPropertyEnvLinkerPassTest.scala
@@ -1,0 +1,7 @@
+package ai.privado.passes.propertyEnv
+
+class JavaPropertyEnvLinkerPassTest {
+
+}
+
+
